Fix fonts not loading after page switch
For some reason, when the page event is triggered, the `head` element returned by `editor.Canvas.getDocument().head` is not the same as in the iframe.

I'm not sure about the root cause yet, it seems like the Canvas is not ready at that time. Might be an issue with GrapesJS, or maybe a wrong assumption in this library?

In any case, a quick workaround is to delay event handling a little.